python/requirements.txt
grpcio
grpcio-tools
tweepy
python-dotenv
